backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
rapidocr-onnxruntime==1.3.15
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
